[LCR-188] refactor: dto명 수정
