+++ src/main/java/com/lgcns/domain/reservation/controller/ReservationController.java
@@ -1,6 +1,6 @@
 package com.lgcns.domain.reservation.controller;
 
-import com.lgcns.domain.reservation.dto.MonthlyReservationResponse;
+import com.lgcns.domain.reservation.dto.response.MonthlyReservationDto;
 import com.lgcns.domain.reservation.service.ReservationService;
 import com.lgcns.global.common.annotation.RawResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -19,7 +19,7 @@ public class ReservationController {
     @GetMapping("/popups/{popupId}")
     @RawResponse
     @Operation(summary = "해당 달에 대한 예약 목록 조회", description = "팝업스토어 ID와 달을 통해 예약 정보를 조회합니다.")
-    public MonthlyReservationResponse userReservationListFind(
+    public MonthlyReservationDto userReservationListFind(
             @PathVariable Long popupId, @RequestParam String date) {
         return reservationService.findReservationByIdAndDate(popupId, date);
     }
