+++ src/main/java/com/lgcns/domain/reservation/dto/response/DailyReservation.java
@@ -1,4 +1,4 @@
-package com.lgcns.domain.reservation.dto;
+package com.lgcns.domain.reservation.dto.response;
 
 import java.time.LocalDate;
 import java.util.List;
