+++ src/main/java/com/lgcns/domain/reservation/repository/ReservationRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.lgcns.domain.reservation.repository;
 
 import com.lgcns.domain.reservation.domain.Reservation;
-import com.lgcns.domain.reservation.dto.DailyReservation;
+import com.lgcns.domain.reservation.dto.response.DailyReservation;
 import java.util.List;
 
 public interface ReservationRepositoryCustom {
